Identity server setup

+++ src/AuctionService/Controllers/AuctionController.cs
@@ -6,6 +6,7 @@ using AutoMapper;
 using AutoMapper.QueryableExtensions;
 using Contracts;
 using MassTransit;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 
@@ -50,13 +51,13 @@ public class AuctionController : ControllerBase
         if (auction == null) return NotFound();
         return _mapper.Map<AuctionDto>(auction);
     }
-
+    [Authorize]
     [HttpPost]
     public async Task<ActionResult<AuctionDto>> CreateAuction(CreateAuctionDto auctionDto)
     {
         var auction = _mapper.Map<Auction>(auctionDto);
         //ToDo: add current user as seller
-        auction.Seller = "test";
+        auction.Seller = User.Identity.Name;
 
         _context.Auctions.Add(auction);
 
@@ -71,7 +72,7 @@ public class AuctionController : ControllerBase
         return CreatedAtAction(nameof(GetAuctionById),
             new { auction.Id }, newAuction);
     }
-
+    [Authorize]
     [HttpPut("{id}")]
     public async Task<ActionResult> UpdateAuction(Guid id, UpdateAuctionDto updateAuctionDto)
     {
@@ -81,7 +82,7 @@ public class AuctionController : ControllerBase
 
         if (auction == null) return NotFound();
 
-        //ToDo: only seller can update auction
+        if (auction.Seller != User.Identity.Name) return Forbid();
 
         auction.Item.Make = updateAuctionDto.Make ?? auction.Item.Make;
         auction.Item.Model = updateAuctionDto.Model ?? auction.Item.Model;
@@ -92,11 +93,12 @@ public class AuctionController : ControllerBase
         await _publishEndpoint.Publish(_mapper.Map<AuctionUpdated>(auction));
 
         var result = await _context.SaveChangesAsync() > 0;
-        
+
         if (!result) return BadRequest("Failed to update auction");
         return Ok();
     }
 
+    [Authorize]
     [HttpDelete("{id}")]
     public async Task<ActionResult> DeleteAuction(Guid id)
     {
@@ -104,7 +106,7 @@ public class AuctionController : ControllerBase
 
         if (auction == null) return NotFound();
 
-        //ToDo: check seller == username
+        if (auction.Seller != User.Identity.Name) return Forbid();
         
         _context.Auctions.Remove(auction);
